+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
   [sgeulette]
 - Added `yesno_value` index to store archived value and possibly use it in a faceted widget.
   [sgeulette]
+- Added `state_filesize` column in `FolderFacetedTableView`
+  [sgeulette]
 
 1.0.2 (2023-11-28)
 ------------------